Fix clippy warnings and suppress intentional dead code
Auto-fixed: unused imports, unnecessary casts, manual prefix stripping,
loop variable indexing, &PathBuf → &Path, &mut Vec → &mut [_].

Suppressed with #[allow(dead_code)]: scaffolded modules (voice, vision,
model-runtime, streaming) that have code written but not yet wired into
the server. These will be activated as features mature.

103 → 53 warnings (remaining are MSRV, too-many-args, method naming —
addressed when CI enforces clippy -D warnings).

+++ myelix-server/src/team_tools.rs
@@ -121,7 +121,7 @@ impl TeamRegistry {
 
         // Check agent limit
         let current = self.total_agents.load(Ordering::Relaxed);
-        if current >= budget.max_agents as u32 {
+        if current >= budget.max_agents {
             return Err(format!(
                 "Maximum agents exceeded ({}/{})",
                 current, budget.max_agents
@@ -173,7 +173,7 @@ impl TeamRegistry {
 
         // Check agent budget
         let current = self.total_agents.load(Ordering::Relaxed);
-        if current >= team.budget.max_agents as u32 {
+        if current >= team.budget.max_agents {
             return Err(format!(
                 "Agent budget exceeded ({}/{})",
                 current, team.budget.max_agents
